BethaX PDF Report done.

+++ grails-app/domain/org/irblleida/bethax/Milestone.groovy
@@ -15,7 +15,7 @@ class Milestone {
     Date deadline
     Date dateFinished
     User headStatistician
-    Float estimatedTime
+    Integer estimatedTime
 
     /** Days to Expiration */
     Integer getDte() {
